fix: optimized sale and product services to use JPQL from repositories

+++ src/main/java/gr/aueb/cf/bluemargarita/dto/sale/WeeklySalesDataDTO.java
@@ -8,7 +8,7 @@ public record WeeklySalesDataDTO(
         int weekOfYear,
         LocalDate weekStartDate,
         LocalDate weekEndDate,
-        BigDecimal quantitySold,
+        Integer quantitySold,
         BigDecimal revenue,
         Integer numberOfSales,
         BigDecimal averagePrice
